👌 Moving current view prompts to front-end. Introducing `aiAssistantHelpers` composable
