+++ packages/editor-ui/src/plugins/i18n/locales/en.json
@@ -150,6 +150,10 @@
 	"aiAssistant.codeUpdated.message.body2": "node to see the changes",
 	"aiAssistant.thinkingSteps.analyzingError": "Analyzing the error...",
 	"aiAssistant.thinkingSteps.thinking": "Thinking...",
+	"aiAssistant.prompts.currentView.workflowList": "The user is currently looking at the list of workflows.",
+	"aiAssistant.prompts.currentView.credentialsList": "The user is currently looking at the list of credentials.",
+	"aiAssistant.prompts.currentView.executionsView": "The user is currently looking at the list of executions for the currently open workflow.",
+	"aiAssistant.prompts.currentView.workflowEditor": "The user is currently looking at the current workflow in n8n editor, without any specific node selected.",
 	"banners.confirmEmail.message.1": "To secure your account and prevent future access issues, please confirm your",
 	"banners.confirmEmail.message.2": "email address.",
 	"banners.confirmEmail.button": "Confirm email",
